+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+* Marker -> View is now allowed to view any question, any paper.
 
 ### Fixed
 * Clients can tag unannotated questions.